fix: apply_chat_template returns BatchEncoding in transformers 5.2
Use return_dict=True and index enc["input_ids"]/enc["attention_mask"] in
TorchPolicy.generate instead of treating the result as a bare tensor.
Also pass attention_mask through to model.generate.

Also fix MPS incompatibility in losses_torch.py: replace .double() with
.float() for clipfrac computation (MPS does not support float64).

+++ pollius/losses_torch.py
@@ -32,7 +32,7 @@ def _ratio(log_prob, old_log_prob):
 def _metrics(loss, ratio, clipped_ratio, neg_approx_kl, mask) -> LossMetrics:
     return LossMetrics(
         loss=float(loss.item()),
-        clipfrac=float(masked_mean((ratio != clipped_ratio).double(), mask).item()),
+        clipfrac=float(masked_mean((ratio != clipped_ratio).float(), mask).item()),
         approx_kl=float(masked_mean(-neg_approx_kl, mask).item()),
         mean_ratio=float(masked_mean(ratio, mask).item()),
     )
